Create API urls and views
