+++ backend/apps/group/urls.py
@@ -28,13 +28,7 @@ urlpatterns = [
     path('', GroupTypeListView.as_view(), name='index'),
     path('<slug:type>/', GroupListView.as_view(), name='sub_index'),
     path('@<slug:slug>', GroupDetailView.as_view(), name='detail'),
-    path('@<slug:slug>/update', GroupDetailView.as_view(), name='update'),
-    path('@<slug:slug>/edit-subscription/', UpdateSubscriptionView.as_view(),
-         name='edit-subscription'),
-    path('@<slug:slug>/edit-membership/', MembershipFormView.as_view(),
-         name='edit-member'),
-    path('@<slug:slug>/admin-request/', AdminRequestFormView.as_view(),
-         name='admin-request')
+    path('@<slug:slug>/edit', GroupDetailView.as_view(), name='update'),
 
     # # edition (réservée aux admins)
     # path(url_base + '<slug:slug>/edit/',
